+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## v0.9.0-rc16 [unreleased]
 
+### Features
+- [#2058](https://github.com/influxdb/influxdb/pull/2058): Track number of queries executed in stats.
+
 ### Bugfixes
 - [#2037](https://github.com/influxdb/influxdb/pull/2037): Don't check 'configExists' at Run() level
 - [#2039](https://github.com/influxdb/influxdb/pull/2039): Don't panic if getting current use fails